Move Error enum to the error module
The trait implementations for the Error enum contain a lot of
boilerplate code and are not very insightful.  Therefore this patch
moves the Error enum, its trait implementations and the related Result
type definition from the main module to the error module.  The Error and
Result types are re-exported from the main module so that the public
interface does not change.