Return Box<dyn Backend> in default_backend()
This patch refactors the default_backend function to return a Box<dyn
Backend> instead of impl Backend.  This will allow us to dynamically
choose the backend implementation in a future patch.  To keep the
current interface, we change show_with to accept both a reference to a
backend instance as well as a reference to a boxed backend instance.
This also means we have to implement AsRef<Self> for the backend
structs.